renamed recipeService to RecipeService updated RecipeService by making it maintainable renamed Recipe to RecipeScreen updated RecipeScreen to use RecipeService renamed the Recipe folder to RecipeInfo
